Fix: bugs

+++ Dataset.py
@@ -24,7 +24,7 @@ class DeblurDataset(Dataset):
 
         line = self.img_path[index]
         img_path, label = line.split('\t')
-        label = np.array([np.double(label)])
+        label = np.array([float(label)])
         label = np.expand_dims(label, axis=0)
 
         img_A = Image.open(img_path + '_blur.png').convert('L')
@@ -48,7 +48,7 @@ class DeblurDataset(Dataset):
         img_A = self.transform(img_A)
         img_B = self.transform(img_B)
 
-        label = transforms.ToTensor()(label)
+        label = transforms.ToTensor()(label).double()
         return img_A, img_B, label, img_name
 
     def __len__(self):
